main/requirements.txt
Flask==2.0.2
Jinja2==3.0.3
json2html==1.3.0
psycopg2-binary==2.9.3
